# Always on tracing with rule counters