--- a/rego/rego.go
+++ b/rego/rego.go
@@ -89,6 +89,7 @@ type EvalContext struct {
 	instrumentation  *topdown.Instrumentation
 	partialNamespace string
 	tracers          []topdown.Tracer
+	queryTracers     []topdown.QueryTracer
 	compiledQuery    compiledQuery
 	unknowns         []string
 	disableInlining  []ast.Ref
@@ -137,10 +138,20 @@ func EvalInstrument(instrument bool) EvalOption {
 }
 
 // EvalTracer configures a tracer for a Prepared Query's evaluation
+// Deprecated: Use EvalQueryTracer instead.
 func EvalTracer(tracer topdown.Tracer) EvalOption {
 	return func(e *EvalContext) {
 		if tracer != nil {
 			e.tracers = append(e.tracers, tracer)
+		}
+	}
+}
+
+// EvalQueryTracer configures a tracer for a Prepared Query's evaluation
+func EvalQueryTracer(tracer topdown.QueryTracer) EvalOption {
+	return func(e *EvalContext) {
+		if tracer != nil {
+			e.queryTracers = append(e.queryTracers, tracer)
 		}
 	}
 }
@@ -217,6 +228,7 @@ func (pq preparedQuery) newEvalContext(ctx context.Context, options []EvalOption
 		instrumentation:  nil,
 		partialNamespace: pq.r.partialNamespace,
 		tracers:          nil,
+		queryTracers:     nil,
 		unknowns:         pq.r.unknowns,
 		parsedUnknowns:   pq.r.parsedUnknowns,
 		compiledQuery:    compiledQuery{},
@@ -457,6 +469,7 @@ type Rego struct {
 	txn                  storage.Transaction
 	metrics              metrics.Metrics
 	tracers              []topdown.Tracer
+	queryTracers         []topdown.QueryTracer
 	tracebuf             *topdown.BufferTracer
 	trace                bool
 	instrumentation      *topdown.Instrumentation
@@ -875,10 +888,20 @@ func Trace(yes bool) func(r *Rego) {
 }
 
 // Tracer returns an argument that adds a query tracer to r.
+// Deprecated: Use QueryTracer instead.
 func Tracer(t topdown.Tracer) func(r *Rego) {
 	return func(r *Rego) {
 		if t != nil {
 			r.tracers = append(r.tracers, t)
+		}
+	}
+}
+
+// QueryTracer returns an argument that adds a query tracer to r.
+func QueryTracer(t topdown.QueryTracer) func(r *Rego) {
+	return func(r *Rego) {
+		if t != nil {
+			r.queryTracers = append(r.queryTracers, t)
 		}
 	}
 }
@@ -992,6 +1015,10 @@ func (r *Rego) Eval(ctx context.Context) (ResultSet, error) {
 
 	for _, t := range r.tracers {
 		evalArgs = append(evalArgs, EvalTracer(t))
+	}
+
+	for _, qt := range r.queryTracers {
+		evalArgs = append(evalArgs, EvalQueryTracer(qt))
 	}
 
 	rs, err := pq.Eval(ctx, evalArgs...)
@@ -1619,6 +1646,10 @@ func (r *Rego) eval(ctx context.Context, ectx *EvalContext) (ResultSet, error) {
 		q = q.WithTracer(ectx.tracers[i])
 	}
 
+	for i := range ectx.queryTracers {
+		q = q.WithQueryTracer(ectx.queryTracers[i])
+	}
+
 	if ectx.parsedInput != nil {
 		q = q.WithInput(ast.NewTerm(ectx.parsedInput))
 	}
@@ -1804,6 +1835,10 @@ func (r *Rego) partial(ctx context.Context, ectx *EvalContext) (*PartialQueries,
 
 	for i := range ectx.tracers {
 		q = q.WithTracer(ectx.tracers[i])
+	}
+
+	for i := range ectx.queryTracers {
+		q = q.WithQueryTracer(ectx.queryTracers[i])
 	}
 
 	if ectx.parsedInput != nil {