--- a/topdown/query.go
+++ b/topdown/query.go
@@ -103,7 +103,7 @@ func (q *Query) WithInput(input *ast.Term) *Query {
 func (q *Query) WithTracer(tracer Tracer) *Query {
 	qt, ok := tracer.(QueryTracer)
 	if !ok {
-		qt = wrapLegacyTracer(tracer)
+		qt = WrapLegacyTracer(tracer)
 	}
 	return q.WithQueryTracer(qt)
 }