At a high level this PR is adding in a new `topdown.QueryTracer` implementation which keeps track of rule's encountered in evaluation and counting their enter/exit/fail events. This information is currently only exposed via the decision logs and evaluations via the REST API. The tracer is always on for any of the REST API's which will perform an evaluation.

The query tracer itself adds a reasonably small amount of overhead during evaluation, scaling with the number of rules involved. That being said, in practice the majority of the impact happens while processing the result in the decision log plugin.

To make this happen there are a number of changes involved, see each commit for more details.

There will still need to be some changes made to expose this via `opa eval` and the repl. Those changes will be done as a follow on for this.