@patrick-east 2020-06-22
Some before and after benchmarks of interest:
```
VirtualDocs1x1-8        10.3µs ± 2%    11.1µs ± 2%   +8.18%  (p=0.000 n=8+10)
VirtualDocs10x1-8       10.4µs ± 2%    11.4µs ± 3%   +8.87%  (p=0.000 n=8+9)
VirtualDocs100x1-8      10.5µs ± 1%    11.7µs ±15%  +12.03%  (p=0.000 n=8+10)
VirtualDocs1000x1-8     10.4µs ± 1%    11.1µs ± 8%   +7.40%  (p=0.000 n=9+10)
VirtualDocs10x10-8      28.9µs ± 1%    32.7µs ± 1%  +13.48%  (p=0.000 n=10+10)
VirtualDocs100x10-8     29.8µs ± 1%    34.1µs ± 4%  +14.66%  (p=0.000 n=10+10)
VirtualDocs1000x10-8    29.6µs ± 2%    33.8µs ± 2%  +14.40%  (p=0.000 n=9+10)

name                  old alloc/op   new alloc/op   delta
VirtualDocs1x1-8        5.52kB ± 0%    5.87kB ± 0%   +6.38%  (p=0.000 n=10+10)
VirtualDocs10x1-8       5.52kB ± 0%    5.87kB ± 0%   +6.38%  (p=0.000 n=10+10)
VirtualDocs100x1-8      5.52kB ± 0%    5.87kB ± 0%   +6.38%  (p=0.000 n=10+10)
VirtualDocs1000x1-8     5.52kB ± 0%    5.87kB ± 0%   +6.38%  (p=0.000 n=10+10)
VirtualDocs10x10-8      19.1kB ± 0%    20.1kB ± 0%   +5.57%  (p=0.000 n=9+10)
VirtualDocs100x10-8     19.1kB ± 0%    20.1kB ± 0%   +5.57%  (p=0.000 n=10+10)
VirtualDocs1000x10-8    19.1kB ± 0%    20.1kB ± 0%   +5.57%  (p=0.000 n=9+8)

name                  old allocs/op  new allocs/op  delta
VirtualDocs1x1-8          96.0 ± 0%      97.0 ± 0%   +1.04%  (p=0.000 n=10+10)
VirtualDocs10x1-8         96.0 ± 0%      97.0 ± 0%   +1.04%  (p=0.000 n=10+10)
VirtualDocs100x1-8        96.0 ± 0%      97.0 ± 0%   +1.04%  (p=0.000 n=10+10)
VirtualDocs1000x1-8       96.0 ± 0%      97.0 ± 0%   +1.04%  (p=0.000 n=10+10)
VirtualDocs10x10-8         273 ± 0%       275 ± 0%   +0.73%  (p=0.000 n=10+10)
VirtualDocs100x10-8        273 ± 0%       275 ± 0%   +0.73%  (p=0.000 n=10+10)
VirtualDocs1000x10-8       273 ± 0%       275 ± 0%   +0.73%  (p=0.000 n=10+10)
```
Note that the comparison above is showing the overhead of _only_ the rule tracer.

When taking into account the full e2e REST API stack w/ decision logging we see changes like:
```
name                                             old time/op    new time/op    delta
pkg:github.com/open-policy-agent/opa/test/e2e/authz goos:darwin goarch:amd64
RESTAuthzForbidAuthn-8                              149µs ± 1%     151µs ± 1%    +1.12%  (p=0.000 n=10+9)
RESTAuthzForbidPath-8                               190µs ± 2%     191µs ± 1%      ~     (p=0.529 n=10+10)
RESTAuthzForbidMethod-8                             201µs ± 2%     199µs ± 2%      ~     (p=0.315 n=10+10)
RESTAuthzAllow10Paths-8                             198µs ± 1%     198µs ± 1%      ~     (p=0.139 n=8+9)
RESTAuthzAllow100Paths-8                            472µs ± 3%     479µs ± 2%    +1.50%  (p=0.035 n=10+9)
RESTAuthzAllow1000Paths-8                          3.21ms ± 2%    3.33ms ± 3%    +3.74%  (p=0.000 n=9+10)
pkg:github.com/open-policy-agent/opa/test/e2e/logs/console goos:darwin goarch:amd64
RESTConsoleDecisionLogger/1x0-8                     224µs ± 0%     206µs ± 0%    -8.00%  (p=0.000 n=8+10)
RESTConsoleDecisionLogger/10x0-8                    224µs ± 1%     206µs ± 1%    -8.23%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/100x0-8                   226µs ± 1%     207µs ± 1%    -8.46%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1000x0-8                  234µs ± 4%     210µs ± 3%    -9.97%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1x1-8                     236µs ± 0%     239µs ± 1%    +1.20%  (p=0.000 n=8+10)
RESTConsoleDecisionLogger/10x1-8                    239µs ± 2%     244µs ± 6%      ~     (p=0.105 n=10+10)
RESTConsoleDecisionLogger/100x1-8                   242µs ± 1%     242µs ± 1%      ~     (p=0.436 n=10+10)
RESTConsoleDecisionLogger/1000x1-8                  245µs ± 1%     247µs ± 2%    +0.97%  (p=0.029 n=10+10)
RESTConsoleDecisionLogger/10x10-8                   252µs ± 2%     299µs ± 0%   +18.62%  (p=0.000 n=10+8)
RESTConsoleDecisionLogger/100x10-8                  255µs ± 2%     301µs ± 1%   +17.98%  (p=0.000 n=10+9)
RESTConsoleDecisionLogger/1000x10-8                 260µs ± 6%     308µs ± 1%   +18.48%  (p=0.000 n=9+10)
RESTConsoleDecisionLogger/100x100-8                 376µs ±13%     835µs ± 0%  +122.16%  (p=0.000 n=10+9)
RESTConsoleDecisionLogger/1000x100-8                414µs ±12%     857µs ± 1%  +107.14%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1000x1000-8              1.90ms ±27%    6.60ms ± 1%  +246.39%  (p=0.000 n=10+10)
pkg:github.com/open-policy-agent/opa/test/e2e/logs/remote goos:darwin goarch:amd64
RESTDecisionLogAuthzForbidAuthn-8                   216µs ± 8%     199µs ± 1%    -7.90%  (p=0.000 n=9+10)
RESTDecisionLogAuthzForbidPath-8                    259µs ±16%     239µs ± 1%    -7.49%  (p=0.000 n=10+10)
RESTDecisionLogAuthzForbidMethod-8                  250µs ± 1%     241µs ± 1%    -3.54%  (p=0.000 n=9+10)
RESTDecisionLogAuthzAllow10Paths-8                  247µs ± 1%     235µs ± 1%    -4.51%  (p=0.000 n=10+10)
RESTDecisionLogAuthzAllow100Paths-8                 525µs ± 1%     521µs ± 3%      ~     (p=0.075 n=10+10)
RESTDecisionLogAuthzAllow1000Paths-8               3.28ms ± 2%    3.35ms ± 2%    +2.14%  (p=0.001 n=10+10)
RESTRemoteDecisionLogger/1x0-8                      202µs ± 3%     181µs ± 4%   -10.63%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/10x0-8                     205µs ± 1%     184µs ± 1%   -10.20%  (p=0.000 n=8+10)
RESTRemoteDecisionLogger/100x0-8                    223µs ±15%     185µs ± 2%   -16.78%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x0-8                   207µs ± 1%     184µs ± 2%   -11.16%  (p=0.000 n=8+10)
RESTRemoteDecisionLogger/1x1-8                      216µs ± 3%     211µs ± 9%      ~     (p=0.243 n=10+9)
RESTRemoteDecisionLogger/10x1-8                     215µs ± 1%     224µs ±11%      ~     (p=1.000 n=9+10)
RESTRemoteDecisionLogger/100x1-8                    215µs ± 2%     218µs ±15%      ~     (p=0.720 n=10+9)
RESTRemoteDecisionLogger/1000x1-8                   215µs ± 2%     208µs ± 3%    -3.42%  (p=0.000 n=10+9)
RESTRemoteDecisionLogger/10x10-8                    228µs ± 2%     246µs ± 9%    +7.85%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/100x10-8                   227µs ± 3%     236µs ± 2%    +3.93%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x10-8                  227µs ± 3%     238µs ± 4%    +4.47%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/100x100-8                  326µs ± 1%     527µs ±13%   +61.66%  (p=0.000 n=8+10)
RESTRemoteDecisionLogger/1000x100-8                 332µs ± 1%     523µs ± 7%   +57.41%  (p=0.000 n=9+9)
RESTRemoteDecisionLogger/1000x1000-8               1.21ms ± 3%    2.60ms ± 5%  +115.61%  (p=0.000 n=10+9)
RESTRemoteDecisionLoggerMaskApplied/1x0-8           283µs ± 1%     269µs ± 7%    -4.72%  (p=0.002 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/10x0-8          283µs ± 1%     264µs ± 1%    -6.76%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/100x0-8         284µs ± 3%     263µs ± 1%    -7.25%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x0-8        285µs ± 5%     263µs ± 1%    -7.62%  (p=0.000 n=9+10)
RESTRemoteDecisionLoggerMaskApplied/1x1-8           326µs ± 9%     279µs ± 2%   -14.45%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/10x1-8          292µs ± 1%     280µs ± 1%    -4.16%  (p=0.000 n=9+10)
RESTRemoteDecisionLoggerMaskApplied/100x1-8         291µs ± 3%     284µs ± 4%    -2.53%  (p=0.013 n=10+9)
RESTRemoteDecisionLoggerMaskApplied/1000x1-8        292µs ± 3%     281µs ± 1%    -3.85%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/10x10-8         304µs ± 2%     317µs ± 2%    +4.06%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/100x10-8        304µs ± 1%     312µs ± 1%    +2.72%  (p=0.000 n=10+9)
RESTRemoteDecisionLoggerMaskApplied/1000x10-8       306µs ± 3%     314µs ± 2%    +2.58%  (p=0.001 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/100x100-8       409µs ± 1%     585µs ± 1%   +43.01%  (p=0.000 n=10+9)
RESTRemoteDecisionLoggerMaskApplied/1000x100-8      410µs ± 2%     585µs ± 1%   +42.48%  (p=0.000 n=9+9)
RESTRemoteDecisionLoggerMaskApplied/1000x1000-8    1.30ms ± 3%    2.65ms ± 2%  +103.58%  (p=0.000 n=10+10)

name                                             old alloc/op   new alloc/op   delta
pkg:github.com/open-policy-agent/opa/test/e2e/authz goos:darwin goarch:amd64
RESTAuthzForbidAuthn-8                             30.5kB ± 0%    31.1kB ± 0%    +1.82%  (p=0.000 n=10+10)
RESTAuthzForbidPath-8                              46.5kB ± 0%    47.0kB ± 0%    +1.18%  (p=0.000 n=10+10)
RESTAuthzForbidMethod-8                            47.1kB ± 0%    47.6kB ± 0%    +1.16%  (p=0.000 n=10+10)
RESTAuthzAllow10Paths-8                            45.8kB ± 0%    46.4kB ± 0%    +1.27%  (p=0.000 n=10+10)
RESTAuthzAllow100Paths-8                            175kB ± 0%     176kB ± 0%    +0.27%  (p=0.000 n=10+10)
RESTAuthzAllow1000Paths-8                          1.47MB ± 0%    1.47MB ± 0%    +0.02%  (p=0.035 n=10+10)
pkg:github.com/open-policy-agent/opa/test/e2e/logs/console goos:darwin goarch:amd64
RESTConsoleDecisionLogger/1x0-8                    52.7kB ± 0%    46.3kB ± 0%   -12.15%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/10x0-8                   52.6kB ± 0%    46.4kB ± 0%   -11.84%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/100x0-8                  52.9kB ± 1%    46.7kB ± 0%   -11.78%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1000x0-8                 52.9kB ± 0%    46.6kB ± 0%   -11.89%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1x1-8                    56.4kB ± 0%    58.8kB ± 0%    +4.33%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/10x1-8                   56.3kB ± 0%    58.8kB ± 0%    +4.43%  (p=0.000 n=10+9)
RESTConsoleDecisionLogger/100x1-8                  56.5kB ± 0%    59.1kB ± 0%    +4.61%  (p=0.000 n=10+9)
RESTConsoleDecisionLogger/1000x1-8                 56.2kB ± 0%    58.8kB ± 0%    +4.68%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/10x10-8                  63.8kB ± 0%    82.6kB ± 0%   +29.40%  (p=0.000 n=9+10)
RESTConsoleDecisionLogger/100x10-8                 64.0kB ± 1%    82.8kB ± 0%   +29.32%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1000x10-8                63.6kB ± 0%    82.3kB ± 0%   +29.44%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/100x100-8                 141kB ± 0%     351kB ± 0%  +148.67%  (p=0.000 n=10+9)
RESTConsoleDecisionLogger/1000x100-8                140kB ± 0%     348kB ± 0%  +148.42%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1000x1000-8               959kB ± 0%    3120kB ± 1%  +225.35%  (p=0.000 n=10+9)
pkg:github.com/open-policy-agent/opa/test/e2e/logs/remote goos:darwin goarch:amd64
RESTDecisionLogAuthzForbidAuthn-8                  57.9kB ± 0%    60.2kB ± 0%    +3.97%  (p=0.000 n=10+10)
RESTDecisionLogAuthzForbidPath-8                   73.5kB ± 0%    75.9kB ± 0%    +3.25%  (p=0.000 n=10+10)
RESTDecisionLogAuthzForbidMethod-8                 74.2kB ± 0%    76.4kB ± 0%    +2.96%  (p=0.000 n=10+10)
RESTDecisionLogAuthzAllow10Paths-8                 73.0kB ± 0%    73.5kB ± 0%    +0.58%  (p=0.000 n=10+10)
RESTDecisionLogAuthzAllow100Paths-8                 203kB ± 0%     203kB ± 0%    +0.25%  (p=0.000 n=10+10)
RESTDecisionLogAuthzAllow1000Paths-8               1.50MB ± 0%    1.50MB ± 0%      ~     (p=0.190 n=10+10)
RESTRemoteDecisionLogger/1x0-8                     52.1kB ± 0%    45.9kB ± 0%   -11.91%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/10x0-8                    52.1kB ± 0%    45.9kB ± 0%   -11.90%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/100x0-8                   52.1kB ± 0%    45.9kB ± 0%   -11.91%  (p=0.000 n=9+10)
RESTRemoteDecisionLogger/1000x0-8                  52.1kB ± 0%    45.9kB ± 0%   -11.86%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1x1-8                     55.4kB ± 0%    55.7kB ± 0%    +0.61%  (p=0.000 n=9+10)
RESTRemoteDecisionLogger/10x1-8                    55.4kB ± 0%    55.7kB ± 0%    +0.54%  (p=0.000 n=10+9)
RESTRemoteDecisionLogger/100x1-8                   55.5kB ± 1%    55.8kB ± 0%    +0.54%  (p=0.001 n=10+9)
RESTRemoteDecisionLogger/1000x1-8                  55.5kB ± 0%    55.8kB ± 0%    +0.55%  (p=0.000 n=10+9)
RESTRemoteDecisionLogger/10x10-8                   62.6kB ± 0%    79.3kB ± 0%   +26.67%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/100x10-8                  62.6kB ± 0%    79.3kB ± 0%   +26.67%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x10-8                 62.6kB ± 0%    79.4kB ± 0%   +26.70%  (p=0.000 n=9+9)
RESTRemoteDecisionLogger/100x100-8                  138kB ± 0%     330kB ± 0%  +139.53%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x100-8                 138kB ± 0%     330kB ± 0%  +139.67%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x1000-8                943kB ± 0%    1451kB ± 0%   +53.84%  (p=0.000 n=9+10)
RESTRemoteDecisionLoggerMaskApplied/1x0-8          95.8kB ± 0%    89.6kB ± 0%    -6.47%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/10x0-8         95.9kB ± 0%    89.6kB ± 0%    -6.49%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/100x0-8        95.9kB ± 0%    89.6kB ± 0%    -6.52%  (p=0.000 n=9+9)
RESTRemoteDecisionLoggerMaskApplied/1000x0-8       95.9kB ± 0%    89.7kB ± 0%    -6.45%  (p=0.000 n=8+10)
RESTRemoteDecisionLoggerMaskApplied/1x1-8          99.0kB ± 0%    99.6kB ± 0%    +0.58%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/10x1-8         99.1kB ± 0%    99.6kB ± 0%    +0.53%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/100x1-8        99.0kB ± 0%    99.6kB ± 0%    +0.52%  (p=0.000 n=9+10)
RESTRemoteDecisionLoggerMaskApplied/1000x1-8       99.1kB ± 0%    99.6kB ± 0%    +0.54%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/10x10-8         106kB ± 0%     123kB ± 0%   +15.90%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/100x10-8        106kB ± 0%     123kB ± 0%   +15.91%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x10-8       106kB ± 0%     123kB ± 0%   +15.93%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/100x100-8       181kB ± 0%     373kB ± 0%  +105.48%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x100-8      181kB ± 0%     373kB ± 0%  +105.39%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x1000-8     987kB ± 0%    1495kB ± 0%   +51.45%  (p=0.000 n=10+10)

name                                             old allocs/op  new allocs/op  delta
pkg:github.com/open-policy-agent/opa/test/e2e/authz goos:darwin goarch:amd64
RESTAuthzForbidAuthn-8                                410 ± 0%       416 ± 0%    +1.46%  (p=0.000 n=10+10)
RESTAuthzForbidPath-8                                 828 ± 0%       834 ± 0%    +0.72%  (p=0.000 n=10+10)
RESTAuthzForbidMethod-8                               837 ± 0%       843 ± 0%    +0.72%  (p=0.000 n=10+10)
RESTAuthzAllow10Paths-8                               821 ± 0%       827 ± 0%    +0.73%  (p=0.000 n=10+10)
RESTAuthzAllow100Paths-8                            4.34k ± 0%     4.35k ± 0%    +0.14%  (p=0.000 n=10+10)
RESTAuthzAllow1000Paths-8                           40.4k ± 0%     40.4k ± 0%    +0.01%  (p=0.000 n=9+10)
pkg:github.com/open-policy-agent/opa/test/e2e/logs/console goos:darwin goarch:amd64
RESTConsoleDecisionLogger/1x0-8                       832 ± 0%       703 ± 0%   -15.55%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/10x0-8                      832 ± 0%       703 ± 0%   -15.50%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/100x0-8                     832 ± 0%       703 ± 0%   -15.50%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1000x0-8                    832 ± 0%       703 ± 0%   -15.50%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1x1-8                       883 ± 0%       830 ± 0%    -6.00%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/10x1-8                      883 ± 0%       830 ± 0%    -6.00%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/100x1-8                     883 ± 0%       830 ± 0%    -6.00%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/1000x1-8                    883 ± 0%       830 ± 0%    -6.00%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/10x10-8                     979 ± 0%      1212 ± 0%   +23.80%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/100x10-8                    980 ± 0%      1212 ± 0%   +23.71%  (p=0.000 n=10+9)
RESTConsoleDecisionLogger/1000x10-8                   979 ± 0%      1212 ± 0%   +23.80%  (p=0.000 n=10+10)
RESTConsoleDecisionLogger/100x100-8                 1.89k ± 0%     4.94k ± 0%  +160.72%  (p=0.000 n=10+9)
RESTConsoleDecisionLogger/1000x100-8                1.89k ± 0%     4.93k ± 0%  +160.70%  (p=0.000 n=10+8)
RESTConsoleDecisionLogger/1000x1000-8               10.9k ± 0%     41.9k ± 0%  +283.61%  (p=0.000 n=7+9)
pkg:github.com/open-policy-agent/opa/test/e2e/logs/remote goos:darwin goarch:amd64
RESTDecisionLogAuthzForbidAuthn-8                     738 ± 0%       629 ± 0%   -14.71%  (p=0.000 n=10+10)
RESTDecisionLogAuthzForbidPath-8                    1.16k ± 0%     1.05k ± 0%    -9.39%  (p=0.000 n=10+10)
RESTDecisionLogAuthzForbidMethod-8                  1.16k ± 0%     1.06k ± 0%    -9.28%  (p=0.000 n=9+9)
RESTDecisionLogAuthzAllow10Paths-8                  1.15k ± 0%     1.03k ± 0%   -10.19%  (p=0.000 n=10+10)
RESTDecisionLogAuthzAllow100Paths-8                 4.67k ± 0%     4.55k ± 0%    -2.50%  (p=0.000 n=10+10)
RESTDecisionLogAuthzAllow1000Paths-8                40.7k ± 0%     40.6k ± 0%    -0.29%  (p=0.000 n=10+9)
RESTRemoteDecisionLogger/1x0-8                        672 ± 0%       543 ± 0%   -19.20%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/10x0-8                       672 ± 0%       543 ± 0%   -19.20%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/100x0-8                      672 ± 0%       543 ± 0%   -19.20%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x0-8                     672 ± 0%       543 ± 0%   -19.20%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1x1-8                        717 ± 0%       603 ± 0%   -15.90%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/10x1-8                       717 ± 0%       603 ± 0%   -15.90%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/100x1-8                      717 ± 0%       603 ± 0%   -15.90%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x1-8                     717 ± 0%       603 ± 0%   -15.90%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/10x10-8                      813 ± 0%       772 ± 0%    -5.04%  (p=0.000 n=10+8)
RESTRemoteDecisionLogger/100x10-8                     813 ± 0%       772 ± 0%    -5.04%  (p=0.000 n=10+9)
RESTRemoteDecisionLogger/1000x10-8                    813 ± 0%       772 ± 0%    -5.01%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/100x100-8                  1.73k ± 0%     2.42k ± 0%   +40.06%  (p=0.000 n=10+10)
RESTRemoteDecisionLogger/1000x100-8                 1.73k ± 0%     2.42k ± 0%   +40.06%  (p=0.000 n=10+9)
RESTRemoteDecisionLogger/1000x1000-8                10.7k ± 0%     18.7k ± 0%   +73.87%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1x0-8           1.05k ± 0%     0.92k ± 0%   -12.30%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/10x0-8          1.05k ± 0%     0.92k ± 0%   -12.30%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/100x0-8         1.05k ± 0%     0.92k ± 0%   -12.30%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x0-8        1.05k ± 0%     0.92k ± 0%   -12.30%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1x1-8           1.09k ± 0%     0.98k ± 0%   -10.42%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/10x1-8          1.09k ± 0%     0.98k ± 0%   -10.42%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/100x1-8         1.09k ± 0%     0.98k ± 0%   -10.42%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/1000x1-8        1.09k ± 0%     0.98k ± 0%   -10.42%  (p=0.000 n=10+10)
RESTRemoteDecisionLoggerMaskApplied/10x10-8         1.19k ± 0%     1.15k ± 0%    -3.36%  (p=0.000 n=10+9)
…